Add r4d support
Add support for r4d (Remote For Device-under-test) which is an
infrastructure for power-control and console access for multiple Linux
boards that should be controlled by a test-infrastructure like Jenkins.
While at it expand information box for hypervisor settings (ssh port) to
r4d needs.

A libvirt version with a patch for the r4d hypervisor is available online.

Link: https://github.com/ci-rt/libvirt-debian
Signed-off-by: Benedikt Spranger <b.spranger@linutronix.de>
Reviewed-by: Bastian Germann <bage@linutronix.de>