JENKINS-50552: Fix triggering snapshot reverts
Configuring a snapshot revert within the project via the "Use Snapshots"
checkbox or via the "Revert Snapshot" and "Before Snapshot" options on the
node configuration does not work. Make the snapshot reverts trigger.

Introduce a ComputerUils class that is called from both RunListener
implementations and makes the LibvirtSnapshotRevertRunListener much simpler,
getting rid of deep if clauses.

[apply some non-functional changes (from warnings), add commit message]
Reviewed-by: Bastian Germann <bage@linutronix.de>