Fix: before Job snapshots for pipelines
This is a fix similar to "JENKINS-50552: Fix triggering snapshot
reverts" but for the 'Before Job' snapshots. It allows to add stages to a
pipeline with an 'agent' directive and have the 'Before Job' snapshot
reverted on the node before each of these stage.

As part of the change I revert "slave: remove circular dependency in
disconnect path" as on our systems it completely broke any job running
on a libvirt node. I hope I fixed the issue this commit was targeting by
removing 'ComputerUtils.disconnect()' from the disconnect method of
VirtualMachineSlaveComputer. This was causing a recursion of disconnects
in my testing. However I can't really be sure it mitigates the original
issue since we never experienced it on our systems. Your testing would
be appreciated.

The main part of the change is to remove
LibvirtSnapshotRevertRunListener and replace it with a new
TaskListener.taskAccepted method in VirtualMachineSlaveComputer.

I also added new OfflineClauses so that when we disconnect to apply a
'Before Job' snapshot we don't also apply a 'Revert' snapshot first if
configured.

Signed-off-by: Michael Jeanson <mjeanson@efficios.com>