Fix: reboot twice after Run
When a node is configured with "Reboot this agent after each build" it
reboots twice at the end of a Run.

This was introduced in commit "JENKINS-50552: Fix triggering snapshot
reverts" which implemented the TaskListener.taskCompleted and
TaskListener.taskCompletedWithProblems methods in
VirtualMachineSlaveComputer that will both reboot the node at the end of
a Task when the node is configured to. The LibvirtRunListener
onFinalized method was kept which similarly reboots the node at the end
of a Run.

The issue is that a Run is composed of a single Task which results in a
reboot at the end of the Task and another reboot at the end of the Run.

Remove the LibvirtRunListener to only reboot at the end of a Task which
covers both pipelines and traditionnal jobs.

Signed-off-by: Michael Jeanson <mjeanson@efficios.com>